Rename `helpers` module to `reader`
