+++ src/parser/mod.rs
@@ -1,8 +1,8 @@
 // Library
 mod bulk_string;
 mod data_types;
-mod helpers;
 mod integer;
+mod reader;
 mod simple_error;
 mod simple_string;
